Delete Account Done

+++ src/app/components/accounts/security/security.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { ToastrService } from 'ngx-toastr';
 import { UsersService } from 'src/app/services/features/users';
 
 @Component({
@@ -9,14 +10,18 @@ import { UsersService } from 'src/app/services/features/users';
 export class SecurityComponent implements OnInit {
 
   constructor(
-    private usersService: UsersService
+    private usersService: UsersService,
+    private toastr: ToastrService
   ) { }
 
   ngOnInit(): void {
   }
 
   async deleteUser(){
-    await this.usersService.deleteUser();
-    await this.usersService.logout();
+    if(confirm('Are you sure you want to delete your account permanently?') == true){
+      await this.usersService.deleteUser();
+      await this.usersService.logout();
+      this.toastr.success('User deleted successfully.', 'Delete Ad!');
+	  }
   }
 }
